Changes tests to desired Tests folder.
