+++ Phantasma.Business/tests/Blockchain/Contracts/Legacy/ExchangeContractTests.cs
@@ -6,16 +6,16 @@ using System.Numerics;
 using Phantasma.Business.Blockchain;
 using Phantasma.Business.Blockchain.Contracts;
 using Phantasma.Business.CodeGen.Assembler;
+using Phantasma.Business.Tests.Simulator;
 using Phantasma.Business.VM.Utils;
 using Phantasma.Core.Cryptography;
 using Phantasma.Core.Domain;
 using Phantasma.Core.Numerics;
-using Phantasma.Simulator;
 
 using Xunit;
 using Shouldly;
 
-namespace Phantasma.Business.Tests.Blockchain.Contracts;
+namespace Phantasma.Business.Tests.Blockchain.Contracts.Legacy;
 
 [Collection("ExchangeContractTests")]
 public class ExchangeContractTests
