+++ Phantasma.Business/tests/Blockchain/Contracts/Legacy/StakeContractTest.cs
@@ -3,16 +3,16 @@ using System.Linq;
 using System.Numerics;
 using Phantasma.Business.Blockchain;
 using Phantasma.Business.Blockchain.Contracts;
+using Phantasma.Business.Tests.Simulator;
 using Phantasma.Business.VM.Utils;
 using Phantasma.Core.Cryptography;
 using Phantasma.Core.Domain;
 using Phantasma.Core.Numerics;
 using Phantasma.Core.Types;
-using Phantasma.Simulator;
 
 using Xunit;
 
-namespace Phantasma.Business.Tests.Blockchain.Contracts;
+namespace Phantasma.Business.Tests.Blockchain.Contracts.Legacy;
 
 
 [Collection("StakeContractTest")]
