+++ Phantasma.Business/tests/Blockchain/Contracts/Legacy/SwapContractTest.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Numerics;
 using Phantasma.Business.Blockchain;
 using Phantasma.Business.Blockchain.Contracts;
+using Phantasma.Business.Tests.Simulator;
 using Phantasma.Business.VM.Utils;
 using Phantasma.Core.Cryptography;
 using Phantasma.Core.Domain;
@@ -10,11 +11,10 @@ using Phantasma.Core.Numerics;
 using Phantasma.Core.Storage.Context;
 using Phantasma.Core.Types;
 using Phantasma.Infrastructure.Pay.Chains;
-using Phantasma.Simulator;
 
 using Xunit;
 
-namespace Phantasma.Business.Tests.Blockchain.Contracts;
+namespace Phantasma.Business.Tests.Blockchain.Contracts.Legacy;
 
 [Collection("SwapContractTest")]
 public class SwapContractTest
