+++ Phantasma.Business/tests/Blockchain/Exploits/Exploits.cs
@@ -1,12 +1,12 @@
 using System;
 using System.Linq;
 using Phantasma.Business.CodeGen.Assembler;
+using Phantasma.Business.Tests.Simulator;
 using Phantasma.Business.VM.Utils;
 using Phantasma.Core.Cryptography;
 using Phantasma.Core.Domain;
 using Phantasma.Core.Numerics;
 using Phantasma.Core.Types;
-using Phantasma.Simulator;
 
 using Xunit;
