+++ Phantasma.Business/tests/Blockchain/Exploits/WhenSystemAddressUsedAsSourceInScript.cs
@@ -1,11 +1,11 @@
 using System.Numerics;
 using Phantasma.Business.Blockchain;
+using Phantasma.Business.Tests.Simulator;
 using Phantasma.Business.VM.Utils;
 using Phantasma.Core.Cryptography;
 using Phantasma.Core.Domain;
 using Phantasma.Core.Numerics;
 using Phantasma.Core.Types;
-using Phantasma.Simulator;
 
 using Xunit;
