+++ Phantasma.Business/tests/Blockchain/NFTTests.cs
@@ -5,12 +5,12 @@ using Xunit;
 using Phantasma.Business.Blockchain;
 using Phantasma.Business.Blockchain.Contracts;
 using Phantasma.Business.Blockchain.Tokens;
+using Phantasma.Business.Tests.Simulator;
 using Phantasma.Business.VM.Utils;
 using Phantasma.Core.Cryptography;
 using Phantasma.Core.Domain;
 using Phantasma.Core.Numerics;
 using Phantasma.Core.Types;
-using Phantasma.Simulator;
 
 namespace Phantasma.Business.Tests.Blockchain;
