+++ Phantasma.Business/tests/Blockchain/Tokens/TokenTests.cs
@@ -3,11 +3,11 @@ using System.Linq;
 using System.Xml.Linq;
 using Phantasma.Business.Blockchain;
 using Phantasma.Business.CodeGen.Assembler;
+using Phantasma.Business.Tests.Simulator;
 using Phantasma.Business.VM.Utils;
 using Phantasma.Core.Cryptography;
 using Phantasma.Core.Domain;
 using Phantasma.Core.Numerics;
-using Phantasma.Simulator;
 
 using Xunit;
