+++ Phantasma.Core/tests/MerkleTreeTests.cs
